scripts: adapt makehordesat.sh
Instead of retrieving MiniSat, get MergeSats git repository, and
build MergeSat instead of MiniSat.

A next step would be to add MergeSat as a git submodule. However, the
required changes for hordesat are not integrated into the production
branch yet, nor in CI. Hence, we will keep this simple for now.

Signed-off-by: Norbert Manthey <nmanthey@conp-solutions.com>