mergesat: do not interrupt search for sharing
The MiniSat backend of hordesat triggered a restart of the solver
by using the interrupt functionality to share new clauses and
assumptions. This results in modifying the original solver
behavior heavily.

This change allows to asynchronously share clauses with the solver,
by providing another hook function which is called when the solver
is ready to receive clauses. By using this function, we allow to
keep the solver behavior and not introduce more restarts.

Signed-off-by: Norbert Manthey <nmanthey@conp-solutions.com>